treewide: init sr models before WAS
WAS needs to be able to decide if an update of the model partition is
required. The easiest way to make this possible is to have Willow send
the supported models in the hello message, so that WAS is aware of them.
Initialize the SR models before initializing WAS so we can add them to
the hello message.