i2c: implement i2c_probe function
And use it to probe for the I2C addresses of touch controllers used in
the ESP32-S3-Box variants supported by Willow. This is in preparation of
switching back to upstream ESP-ADF instead of our fork.

The i2c_bus_probe_addr function in ESP-ADF's i2c_bus driver rejects
addresses that use more than 7 bits, but does not set bit 8 to indicate
that we want to write from master to slave.